requirements.txt
python-dotenv
psycopg2-binary
fastapi
uvicorn
python-multipart
requests
pandas
openpyxl
pymupdf
groq